refactor org.apache.shardingsphere.driver.spring.namespace.constants
