+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/java/org/apache/shardingsphere/driver/spring/namespace/constants/rules/sharding/ShardingStrategyBeanDefinitionParserTag.java
@@ -15,18 +15,28 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.driver.spring.namespace.constants;
+package org.apache.shardingsphere.driver.spring.namespace.constants.rules.sharding;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 /**
- * Master-slave rule parser tag constants.
+ * Sharding strategy parser tag constants.
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class MasterSlaveRuleBeanDefinitionParserTag {
+public final class ShardingStrategyBeanDefinitionParserTag {
     
-    public static final String MASTER_SLAVE_RULE_TAG = "master-slave-rule";
+    public static final String STANDARD_STRATEGY_ROOT_TAG = "standard-strategy";
     
-    public static final String MASTER_SLAVE_DATA_SOURCE_TAG = "data-source-rule";
+    public static final String COMPLEX_STRATEGY_ROOT_TAG = "complex-strategy";
+    
+    public static final String HINT_STRATEGY_ROOT_TAG = "hint-strategy";
+    
+    public static final String NONE_STRATEGY_ROOT_TAG = "none-strategy";
+    
+    public static final String SHARDING_COLUMN_ATTRIBUTE = "sharding-column";
+    
+    public static final String SHARDING_COLUMNS_ATTRIBUTE = "sharding-columns";
+    
+    public static final String ALGORITHM_REF_ATTRIBUTE = "algorithm-ref";
 }
