+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/java/org/apache/shardingsphere/driver/spring/namespace/handler/DataSourceNamespaceHandler.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.driver.spring.namespace.handler;
 
 import org.apache.shardingsphere.driver.spring.namespace.constants.DataSourceBeanDefinitionParserTag;
-import org.apache.shardingsphere.driver.spring.namespace.constants.ShardingTransactionTypeScannerBeanDefinitionParserTag;
+import org.apache.shardingsphere.driver.spring.namespace.constants.rules.sharding.ShardingTransactionTypeScannerBeanDefinitionParserTag;
 import org.apache.shardingsphere.driver.spring.namespace.parser.DataSourceBeanDefinitionParser;
 import org.apache.shardingsphere.driver.spring.namespace.parser.rule.sharding.ShardingTransactionTypeScannerBeanDefinitionParser;
 import org.springframework.beans.factory.xml.NamespaceHandlerSupport;
