+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/java/org/apache/shardingsphere/driver/spring/namespace/parser/rule/sharding/ShardingAlgorithmBeanDefinitionParser.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.driver.spring.namespace.parser.rule.sharding;
 
 import com.google.common.base.Strings;
-import org.apache.shardingsphere.driver.spring.namespace.constants.SPIBeanDefinitionParserTag;
+import org.apache.shardingsphere.driver.spring.namespace.constants.rules.sharding.SPIBeanDefinitionParserTag;
 import org.apache.shardingsphere.driver.spring.namespace.factorybean.ShardingAlgorithmFactoryBean;
 import org.springframework.beans.factory.support.AbstractBeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
